+++ core-ng-test/src/main/java/core/framework/test/kafka/MockKafka.java
@@ -30,7 +30,7 @@ public class MockKafka extends Kafka {
     }
 
     @Override
-    public Consumer<String, byte[]> consumer(String group, Set<String> topics) {
+    public Consumer<String, byte[]> consumer(String group, Set<String> topics, String thread) {
         MockConsumer<String, byte[]> consumer = new MockConsumer<>(OffsetResetStrategy.EARLIEST);
 
         List<TopicPartition> assignments = Lists.newArrayList();