+++ core-ng/src/main/java/core/framework/impl/web/management/UpdateTopicRequest.java
@@ -1,6 +1,11 @@
 package core.framework.impl.web.management;
 
 import core.framework.api.json.Property;
+import core.framework.api.validate.Min;
+import core.framework.api.validate.NotNull;
+import core.framework.api.validate.Size;
+
+import java.util.List;
 
 /**
  * @author neo
@@ -8,4 +13,20 @@ import core.framework.api.json.Property;
 public class UpdateTopicRequest {
     @Property(name = "partitions")
     public Integer partitions;
+
+    @Size(min = 1)
+    @Property(name = "delete_records")
+    public List<DeleteRecord> deleteRecords;
+
+    public static class DeleteRecord {
+        @NotNull
+        @Min(0)
+        @Property(name = "partition")
+        public Integer partition;
+
+        @NotNull
+        @Min(0)
+        @Property(name = "before_offset")
+        public Long beforeOffset;
+    }
 }